svg icon for the French flag
If it may help, here's a svg icon for the French flag using the official colors (as per http://francophilia.com/gazette/?p=3531).  I took the German flag in flags.svg, rotated it, changed colors (and got rid of the transparency ellipse on the black stripe; put it back if you want).